+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 =================
 
 ## 0.18.63
+### is.codion.common.reactive
+- AbstractValue.setAndNotify() bug fixed, change event no longer triggered when nullValue set.
 ### is.codion.swing.common.model
 - DefaultListSelection.DefaultItems.getValue() bug fixed, now excludes invalid indexes.
 ### is.codion.swing.common.ui